Fixed: FS#1808 - Some small bugs in statistics and client selector.
git-svn-id: svn://svn.ispconfig.org/ispconfig3/branches/ispconfig-3.0.4@2735 c1a6f2c7-8504-0410-933a-a4dae0184b0a

+++ server/cron_daily.php
@@ -133,7 +133,7 @@ foreach($records as $rec) {
 	$webalizer_conf_main = '/etc/webalizer/webalizer.conf';
 	$webalizer_conf = escapeshellcmd($rec['document_root'].'/log/webalizer.conf');
 	
-	unlink($statsdir.'/index.php');
+	if(is_file($statsdir.'/index.php')) unlink($statsdir.'/index.php');
 
 	if(!@is_file($webalizer_conf)) {
 		copy($webalizer_conf_main,$webalizer_conf);
